# payment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Gest Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/payment-intents/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-intents"
                ],
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model.PaymentIntent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Bs": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Payment": {
            "type": "object",
            "properties": {
                "bs": {
                    "$ref": "#/definitions/model.Bs"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}